fix(parser): fixed isUrl function and parsing logic
