+++ packages/ng-openapi/src/lib/cli.ts
@@ -6,6 +6,7 @@ import * as fs from "fs";
 import { generateFromConfig } from "./core";
 import * as packageJson from "../../package.json";
 import { GeneratorConfig } from "@ng-openapi/shared";
+import { isUrl } from "@ng-openapi/shared/src/utils/functions/is-url";
 
 const program = new Command();
 
@@ -54,70 +55,12 @@ async function loadConfigFile(configPath: string): Promise<GeneratorConfig> {
     }
 }
 
-function validateInputExtension(inputPath: string): void {
-    const extension = path.extname(inputPath).toLowerCase();
-    const supportedExtensions = [".json", ".yaml", ".yml"];
-
-    if (!supportedExtensions.includes(extension)) {
-        throw new Error(
-            `Failed to parse ${extension || "specification"}. Supported formats are .json, .yaml, and .yml.`
-        );
-    }
-}
-
-function isUrl(input: string): boolean {
-    try {
-        new URL(input);
-        return true;
-    } catch {
-        return false;
-    }
-}
-
-function validateInput(inputPath: string): void {
-    if (isUrl(inputPath)) {
-        // For URLs, we can't validate existence beforehand, but we can validate the URL format
-        const url = new URL(inputPath);
-        if (!["http:", "https:"].includes(url.protocol)) {
-            throw new Error(`Unsupported URL protocol: ${url.protocol}. Only HTTP and HTTPS are supported.`);
-        }
-        return; // URL validation passed
-    }
-
-    // For local files, check existence and extension
-    if (!fs.existsSync(inputPath)) {
-        throw new Error(`Input file not found: ${inputPath}`);
-    }
-
-    const extension = path.extname(inputPath).toLowerCase();
-    const supportedExtensions = [".json", ".yaml", ".yml"];
-
-    if (!supportedExtensions.includes(extension)) {
-        throw new Error(
-            `Failed to parse ${extension || "specification"}. Supported formats are .json, .yaml, and .yml.`
-        );
-    }
-}
 async function generateFromOptions(options: any): Promise<void> {
     try {
         if (options.config) {
-            // Load configuration from file (paths are now resolved in loadConfigFile)
             const config = await loadConfigFile(options.config);
-
-            // Validate the input from config (file or URL)
-            // Only validate local files, not URLs
-            if (!isUrl(config.input)) {
-                if (!fs.existsSync(config.input)) {
-                    throw new Error(`Input file not found: ${config.input}`);
-                }
-                validateInputExtension(config.input);
-            }
-
             await generateFromConfig(config);
         } else if (options.input) {
-            // Use command line options
-            validateInput(options.input);
-
             const config: GeneratorConfig = {
                 input: options.input, // Can now be a URL or file path
                 output: options.output || "./src/generated",
