+++ packages/shared/src/utils/functions/is-url.ts
@@ -1,0 +1,11 @@
+/**
+ * Determines if input is a URL
+ */
+export function isUrl(input: string): boolean {
+    try {
+        const url = new URL(input);
+        return ["http:", "https:"].includes(url.protocol);
+    } catch {
+        return false;
+    }
+}
